requirements.txt
shade
python-keystoneclient
python-monascaclient

